Move job logic to the validator
